refactor:Membership && PointEntity/#10
