+++ src/main/java/com/example/a3sproject/domain/point/entity/PointTransaction.java
@@ -2,6 +2,8 @@ package com.example.a3sproject.domain.point.entity;
 
 import com.example.a3sproject.domain.point.enums.PointTransactionType;
 import com.example.a3sproject.global.entity.BaseEntity;
+import com.example.a3sproject.global.exception.common.ErrorCode;
+import com.example.a3sproject.global.exception.domain.PointException;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -13,7 +15,7 @@ import java.time.LocalDateTime;
 @Getter
 @Table(name = "points")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Point extends BaseEntity {
+public class PointTransaction extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -31,6 +33,10 @@ public class Point extends BaseEntity {
     @Column(nullable = false)
     private int points;
 
+    // 포인트 잔액 (스냅샷)
+    @Column(nullable = false)
+    private int pointBalance;
+
     // 거래 타입
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
@@ -41,15 +47,33 @@ public class Point extends BaseEntity {
     private LocalDateTime expiredAt;
 
     // 정적 팩토리 메서드
-    public static Point of(Long userId, Long orderId,
-                           int points, PointTransactionType type,
-                           LocalDateTime expiredAt) {
-        Point tx = new Point();
+    public static PointTransaction of(Long userId, Long orderId,
+                                      int points, PointTransactionType type,
+                                      LocalDateTime expiredAt) {
+        PointTransaction tx = new PointTransaction();
         tx.userId = userId;
         tx.orderId = orderId;
         tx.points = points;
         tx.type = type;
         tx.expiredAt = expiredAt;
         return tx;
+    }
+
+    // 포인트 차감
+    public void usePoint(int amount) {
+        if (this.pointBalance < amount) {
+            throw new PointException(ErrorCode.POINT_NOT_ENOUGH);
+        }
+        this.pointBalance -= amount;
+    }
+
+    // 포인트 적립
+    public void earnPoint(int amount) {
+        this.pointBalance += amount;
+    }
+
+    // 포인트 복구
+    public void restorePoint(int amount) {
+        this.pointBalance += amount;
     }
 }
