+++ src/main/java/com/example/a3sproject/domain/user/entity/User.java
@@ -33,11 +33,16 @@ public class User extends BaseEntity {
     @Column(nullable = false)
     private String phoneNumber;
 
+    // 포인트 잔액
+    @Column(nullable = false)
+    private Integer pointBalance;
+
     // 회원가입용 생성자
-    public User(String name, String email, String password, String phoneNumber) {
+    public User(String name, String email, String password, String phoneNumber, Integer pointBalance ) {
         this.name = name;
         this.email = email;
         this.password = password;
         this.phoneNumber = phoneNumber;
+        this.pointBalance = pointBalance;
     }
 }
